refactor(skill): align runtime audit boundaries
